+++ module/data/items/item-skill.mjs
@@ -1,5 +1,6 @@
 import SWNItemBase from './base-item.mjs';
 import SWNShared from '../shared.mjs';
+import { getChatMessageMode } from '../../helpers/utils.mjs';
 
 export default class SWNSkill extends SWNItemBase {
   static LOCALIZATION_PREFIXES = [
@@ -58,7 +59,7 @@ export default class SWNSkill extends SWNItemBase {
     modifier,
     unskilledPenaltyMod = 0
   ) {
-    const rollMode = game.settings.get("core", "rollMode");
+    const rollMode = getChatMessageMode();
 
     let formula = `${dice} + @stat + @skill + @modifier`;
     if (skillRank < 0 && game.settings.get("swnr", "unskilledPenalty") != -1) {
@@ -145,7 +146,7 @@ export default class SWNSkill extends SWNItemBase {
       stats: actor.system.stats
     };
 
-    const content = await renderTemplate(template, dialogData);
+    const content = await foundry.applications.handlebars.renderTemplate(template, dialogData);
     const _doRoll = async (_event, button, html) => {
       const dice = button.form.elements.dicepool.value;
       const statShortNameForm = button.form.elements.stat.value;